requirements.txt
langchain==1.2.10
langchain-core==1.2.17
langchain-community==0.4.1
langchain-groq==1.1.2
langchain-huggingface
langchain-classic
faiss-cpu
pypdf
streamlit
python-dotenv
tiktoken
sentence-transformers